bin/requirements.txt
matrix-nio
requests
markdown
asyncio
datetime
json